Simplify declarations struct
